added comment to the function

+++ 0x00-python_variable_annotations/2-floor.py
@@ -1,11 +1,11 @@
 #!/usr/bin/env python3
-import math
 """
      a type-annotated function floor which takes a
      float n as argument and returns the floor of the float.
 """
+import math
 
 
-def floor(n: float) -> float:
+def floor(n: float) -> int:
     """Returns the floor of the float n"""
     return math.floor(n)
